Separate sysrq-trigger test from proc file api tests
Because the sysrq-trigger test reboots the device, if the device does
not boot before the timeout, the failure can cascade and break other
tests in proc_file_api_tests.  Split the test off so that this does not
happen.  Tradefed will wait for the device to become available again
before running proc_file_api_tests.

Bug: 293215373
Test: atest vts_kernel_proc_sysrq_trigger_test
Test: atest vts_kernel_proc_file_api_test
Change-Id: I609cebf6b0484f6d03dae04bf6f6ec586ab60a2e
Signed-off-by: Edward Liaw <edliaw@google.com>