Use the new ro.vendor.api_level for the test screening
In 24Q2, vendor api level was frozen with the api level 202404.
Use this version instead of SDK api level.
ro.vendor.api_level already provides the minimum of the vendor api
level and the product first api level. Read the property to get the
api level for the test screening.
Some tests are required to not only the launching devices but also to
upgrading devices. Read the PropertyUtil.getVendorApiLevel() in this
case.

Bug: 341608258
Test: atest VtsGpuTests
Change-Id: Id24179f6f9f3208362ed0b463844e759c18d3099